incompativeis g

+++ src/MergulhadorJAVA/PromocaoAritmetica.java
@@ -4,7 +4,7 @@ public class PromocaoAritmetica {
     public static void main(String[] args) {
         long x =10;
         long y = 5;
-        int z = x * y;
+        int z = (long) x * y;
 
         System.out.println(z);
     }
